refactor: decompose work_queue.rs into focused submodules
Splits the 1,458-line crates/engine/src/concurrent_delta/work_queue.rs
into a directory with five focused submodules and a sibling tests file.
Public API surface is unchanged - all symbols are re-exported from
work_queue/mod.rs so external callers continue to use the same paths.

Submodule layout:
- bounded.rs (104) - WorkQueueSender / WorkQueueReceiver / SendError /
  bounded() / bounded_with_capacity()
- drain.rs (156) - drain_parallel() / drain_parallel_into() with the
  per-thread sharding logic
- capacity.rs (76) - default_capacity() / adaptive_queue_depth() and
  capacity multiplier constants
- iter.rs (35) - WorkQueueIter and IntoIterator impl
- multi_producer.rs (23) - Clone impl gated on the multi-producer feature
- tests.rs (1,005) - all unit and proptest cases moved out of the
  production source

Tasks #1856 and #1857.
